refactor(workers): import stableWorkerPathComponent from its defining module
workspace-sync.ts crossed the 700-line lint budget by one; drop its
re-export and point consumers at workspace-sync-helpers directly.
