+++ src/gateway/worker-environments/workspace-sync-tunnel.test.ts
@@ -26,7 +26,7 @@ import {
 } from "./tunnel.test-support.js";
 import { rsyncArgvPort, sshArgvPort } from "./worker-ssh-argv.test-support.js";
 import { parseWorkerWorkspaceManifest } from "./workspace-reconcile.js";
-import { stableWorkerPathComponent } from "./workspace-sync.js";
+import { stableWorkerPathComponent } from "./workspace-sync-helpers.js";
 
 const tempDirs = useAutoCleanupTempDirTracker(afterEach);
